+++ db/pvault.sql
@@ -1,6 +1,6 @@
 SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
 
-CREATE TABLE `allergens` (
+CREATE TABLE `ingredient_compounds` (
   `id` int(11) NOT NULL,
   `ing` varchar(255) COLLATE utf8_general_ci NOT NULL,
   `name` varchar(255) COLLATE utf8_general_ci NOT NULL,
@@ -446,7 +446,7 @@ ALTER TABLE `users`
   ADD PRIMARY KEY (`id`);
 
 
-ALTER TABLE `allergens`
+ALTER TABLE `ingredient_compounds`
   ADD PRIMARY KEY (`id`);
 
 ALTER TABLE `bottles`
@@ -491,7 +491,7 @@ ALTER TABLE `settings`
 ALTER TABLE `users`
   MODIFY `id` int(11) NOT NULL AUTO_INCREMENT;
 
-ALTER TABLE `allergens`
+ALTER TABLE `ingredient_compounds`
   MODIFY `id` int(11) NOT NULL AUTO_INCREMENT;
 
 CREATE TABLE `documents` (