+++ db/updates/update_10.1-10.2.sql
@@ -1,0 +1,1 @@
+RENAME TABLE `allergens` TO `ingredient_compounds`;